requirements.txt
aniso8601==5.1.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.43.0
dash-core-components==0.48.0
dash-html-components==0.16.0
dash-renderer==0.24.0
dash-table==3.7.0
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
django-graphql-jwt==0.2.1
django-plotly-dash==0.9.12
django-reversion==3.0.4
django-summernote==0.8.11.4
django-tinymce4-lite==1.7.5
docopt==0.6.2
dpd-components==0.1.0
Flask==1.0.3
Flask-Compress==1.4.0
graphene==2.1.5
graphene-django==2.3.0
graphql-core==2.2
graphql-relay==0.4.5
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
jsonschema==3.0.1
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
Pillow==6.0.0
pipreqs==0.4.9
plotly==3.10.0
promise==2.2.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
relativetimebuilder==1.0.0
requests==2.22.0
retrying==1.3.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
Werkzeug==0.15.4
yarg==0.1.9
